Move android specific classes into the corresponding parent package
