+++ src/main/java/io/appium/java_client/android/appmanagement/AndroidTerminateApplicationOptions.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package io.appium.java_client.appmanagement;
+package io.appium.java_client.android.appmanagement;
 
 import com.google.common.collect.ImmutableMap;
+import io.appium.java_client.appmanagement.BaseTerminateApplicationOptions;
 
 import java.time.Duration;
 import java.util.Map;
@@ -25,43 +26,20 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static java.util.Optional.ofNullable;
 
-public class AndroidRemoveApplicationOptions extends
-        BaseRemoveApplicationOptions<AndroidRemoveApplicationOptions> {
+public class AndroidTerminateApplicationOptions extends
+        BaseTerminateApplicationOptions<AndroidTerminateApplicationOptions> {
     private Duration timeout;
-    private Boolean keepData;
 
     /**
-     * The time to wait until the app is removed (20000ms by default).
+     * The time to wait until the app is terminated (500ms by default).
      *
      * @param timeout the actual timeout value. The minimum time resolution
      *                unit is one millisecond.
      * @return self instance for chaining.
      */
-    public AndroidRemoveApplicationOptions withTimeout(Duration timeout) {
-        checkArgument(!checkNotNull(timeout).isNegative(),
-                "The timeout value cannot be negative");
+    public AndroidTerminateApplicationOptions withTimeout(Duration timeout) {
+        checkArgument(!checkNotNull(timeout).isNegative(), "The timeout value cannot be negative");
         this.timeout = timeout;
-        return this;
-    }
-
-    /**
-     * Forces uninstall to keep the application data and caches.
-     *
-     * @return self instance for chaining.
-     */
-    public AndroidRemoveApplicationOptions withKeepDataEnabled() {
-        this.keepData = true;
-        return this;
-    }
-
-    /**
-     * Forces uninstall to delete the application data and caches
-     * (the default behavior).
-     *
-     * @return self instance for chaining.
-     */
-    public AndroidRemoveApplicationOptions withKeepDataDisabled() {
-        this.keepData = false;
         return this;
     }
 
@@ -70,7 +48,6 @@ public class AndroidRemoveApplicationOptions extends
     public Map<String, Object> build() {
         final ImmutableMap.Builder builder = new ImmutableMap.Builder<String, Object>();
         ofNullable(timeout).map(x -> builder.put("timeout", x.toMillis()));
-        ofNullable(keepData).map(x -> builder.put("keepData", x));
         return builder.build();
     }
 }
